[OTE-TEST] Disable obsolete test cases (#1220)
* Remove (Change to skip) some model template test cases for OTE-CI test (some external backend's API & performance tests)

* Fixed some library version conflict issue (in OTE CI venv)

* Fixed some test cases not working (MPA ote_cli tests)

* Add MPA pre-commit tests (not for pydocstyle -> will check it out later)

* Resolve prettier issue (data/car_tree_bug/annotations/hierarchical_default.json)

* Resolve black issue (ote_cli/ote_cli/utils/hpo.py)
