+++ .pre-commit-config.yaml
@@ -13,7 +13,7 @@ repos:
       - id: isort
         alias: isort_rest
         name: "isort (ote_cli|external)"
-        files: '^(ote_cli|external/anomaly)/.*\.py'
+        files: '^(ote_cli|external/anomaly|external/model-preparation-algorithm)/.*\.py'
         exclude: "tests/"
 
   - repo: https://github.com/psf/black
@@ -26,14 +26,14 @@ repos:
       - id: black
         name: "black (rest)"
         args: [--line-length, "120"]
-        files: '^external/anomaly/.*\.py'
+        files: '^(external/anomaly|external/model-preparation-algorithm)/.*\.py'
 
   - repo: https://github.com/PyCQA/flake8
     rev: "5.0.3"
     hooks:
       - id: flake8
         name: "flake8"
-        files: '^(ote_sdk|ote_cli|external/anomaly)/.*\.py'
+        files: '^(ote_sdk|ote_cli|external/anomaly|external/model-preparation-algorithm)/.*\.py'
         args: ["--config", ".flake8", "--max-complexity", "20"]
         exclude: ".*/protobuf"
 
@@ -70,16 +70,16 @@ repos:
       - id: mypy
         alias: mypy_external
         name: "mypy (external)"
-        files: '^external/anomaly/.*\.py'
+        files: '^(external/anomaly|external/model-preparation-algorithm)/.*\.py'
         additional_dependencies: [attrs==21.2.*]
-        exclude: "^external/anomaly/tests"
+        exclude: "^(external/anomaly|external/model-preparation-algorithm)/tests"
 
   - repo: https://github.com/PyCQA/pylint
     rev: "v2.14.5"
     hooks:
       - id: pylint
         name: "pylint"
-        files: '^(ote_sdk|ote_cli|external/anomaly)/.*\.py'
+        files: '^(ote_sdk|ote_cli|external/anomaly|external/model-preparation-algorithm)/.*\.py'
         entry: pylint
         language: system
         types: [python]
