+++ external/deep-object-reid/constraints.txt
@@ -1,1 +1,2 @@
 opencv-python==4.5.5.64  # remedy for fixed opencv-python-headless version in e2e-test-framework
+optuna==2.10.1 # remedy for fixed optuna version incompatible in OTE CI
