+++ external/mmsegmentation/tests/ote_cli/test_segmentation.py
@@ -72,6 +72,7 @@ templates = Registry('external/mmsegmentation', experimental=True).filter(task_t
 templates_ids = [template.model_template_id for template in templates]
 
 
+@pytest.mark.skip(reason="This test case will be deprecated soon")
 class TestToolsSegmentation:
     @e2e_pytest_component
     def test_create_venv(self):
