+++ external/mmsegmentation/tests/test_ote_training.py
@@ -99,20 +99,20 @@ class SegmentationTrainingTestParameters(DefaultOTETestCreationParametersInterfa
         test_bunches = [
                 dict(
                     model_name=[
-                       'Custom_Semantic_Segmentation_Lite-HRNet-s-mod2_OCR',
-                       'Custom_Semantic_Segmentation_Lite-HRNet-18_OCR',
-                       'Custom_Semantic_Segmentation_Lite-HRNet-18-mod2_OCR',
-                       'Custom_Semantic_Segmentation_Lite-HRNet-x-mod3_OCR',
+                        'Custom_Semantic_Segmentation_Lite-HRNet-s-mod2_OCR',
+                        'Custom_Semantic_Segmentation_Lite-HRNet-18_OCR',
+                        'Custom_Semantic_Segmentation_Lite-HRNet-18-mod2_OCR',
+                        'Custom_Semantic_Segmentation_Lite-HRNet-x-mod3_OCR',
                     ],
                     dataset_name='kvasir_seg_shortened',
                     usecase='precommit',
                 ),
                 dict(
                     model_name=[
-                       'Custom_Semantic_Segmentation_Lite-HRNet-s-mod2_OCR',
-                       'Custom_Semantic_Segmentation_Lite-HRNet-18_OCR',
-                       'Custom_Semantic_Segmentation_Lite-HRNet-18-mod2_OCR',
-                       'Custom_Semantic_Segmentation_Lite-HRNet-x-mod3_OCR',
+                        'Custom_Semantic_Segmentation_Lite-HRNet-s-mod2_OCR',
+                        'Custom_Semantic_Segmentation_Lite-HRNet-18_OCR',
+                        'Custom_Semantic_Segmentation_Lite-HRNet-18-mod2_OCR',
+                        'Custom_Semantic_Segmentation_Lite-HRNet-x-mod3_OCR',
                     ],
                     dataset_name='kvasir_seg',
                     num_training_iters=KEEP_CONFIG_FIELD_VALUE,
@@ -272,6 +272,7 @@ class TestOTEReallifeSegmentation(OTETrainingTestInterface):
         logger.info('data_collector is released')
 
     @e2e_pytest_performance
+    @pytest.mark.skip(reason="This test case will be deprecated soon")
     def test(self,
              test_parameters,
              test_case_fx, data_collector_fx,
