+++ external/model-preparation-algorithm/ote_tests_pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-python_files = test_ote_*.py
+python_files = test_ote_*_api.py test_ote_*_training.py
