+++ external/model-preparation-algorithm/tests/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (C) 2022 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+#
