+++ external/model-preparation-algorithm/tests/config.py
@@ -6,6 +6,9 @@ try:
     import os
     from e2e import config as config_e2e
 
-    config_e2e.repository_name = os.environ.get('TT_REPOSITORY_NAME', 'ote/training_extensions/external/model-preparation-algorithm')
+    config_e2e.repository_name = os.environ.get(
+        "TT_REPOSITORY_NAME",
+        "ote/training_extensions/external/model-preparation-algorithm",
+    )
 except ImportError:
     pass
