+++ external/model-preparation-algorithm/tests/conftest.py
@@ -15,48 +15,63 @@ except ImportError:
     pass
 import config
 import pytest
-from ote_sdk.test_suite.pytest_insertions import *
+from ote_sdk.test_suite.pytest_insertions import (
+    get_pytest_plugins_from_ote,
+    ote_conftest_insertion,
+    ote_pytest_generate_tests_insertion,
+    ote_pytest_addoption_insertion,
+)
 from ote_sdk.test_suite.training_tests_common import REALLIFE_USECASE_CONSTANT
 
 pytest_plugins = get_pytest_plugins_from_ote()
 
-ote_conftest_insertion(default_repository_name='ote/training_extensions/external/model-preparation-algorithm')
+ote_conftest_insertion(
+    default_repository_name="ote/training_extensions/external/model-preparation-algorithm"
+)
+
 
 @pytest.fixture
 def ote_test_domain_fx():
-    return 'model-preparation-algorithm'
+    raise NotImplementedError("Please, implement the fixture ote_test_domain_fx in your test file")
+
 
 @pytest.fixture
 def ote_test_scenario_fx(current_test_parameters_fx):
     assert isinstance(current_test_parameters_fx, dict)
-    if current_test_parameters_fx.get('usecase') == REALLIFE_USECASE_CONSTANT:
-        return 'performance'
+    if current_test_parameters_fx.get("usecase") == REALLIFE_USECASE_CONSTANT:
+        return "performance"
     else:
-        return 'integration'
+        return "integration"
 
-@pytest.fixture(scope='session')
+
+@pytest.fixture(scope="session")
 def ote_templates_root_dir_fx():
     import os.path as osp
     import logging
+
     logger = logging.getLogger(__name__)
     root = osp.dirname(osp.dirname(osp.realpath(__file__)))
-    root = f'{root}/configs/'
-    logger.debug(f'overloaded ote_templates_root_dir_fx: return {root}')
+    root = f"{root}/configs/"
+    logger.debug(f"overloaded ote_templates_root_dir_fx: return {root}")
     return root
 
-@pytest.fixture(scope='session')
+
+@pytest.fixture(scope="session")
 def ote_reference_root_dir_fx():
     import os.path as osp
     import logging
+
     logger = logging.getLogger(__name__)
     root = osp.dirname(osp.dirname(osp.realpath(__file__)))
-    root = f'{root}/tests/reference/'
-    logger.debug(f'overloaded ote_reference_root_dir_fx: return {root}')
+    root = f"{root}/tests/reference/"
+    logger.debug(f"overloaded ote_reference_root_dir_fx: return {root}")
     return root
+
 
 # pytest magic
 def pytest_generate_tests(metafunc):
     ote_pytest_generate_tests_insertion(metafunc)
+
 
 def pytest_addoption(parser):
     ote_pytest_addoption_insertion(parser)
