+++ external/model-preparation-algorithm/tests/ote_cli/test_instance_segmentation.py
@@ -1,4 +1,4 @@
-"""Tests for MPA Class-Incremental Learning for image classification with OTE CLI"""
+"""Tests for MPA Class-Incremental Learning for instance segmentation with OTE CLI"""
 # Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 #
@@ -19,25 +19,23 @@ from ote_cli.utils.tests import (
     ote_eval_deployment_testing,
     ote_eval_openvino_testing,
     ote_eval_testing,
-    ote_hpo_testing,
     ote_train_testing,
     ote_export_testing,
-    pot_optimize_testing,
-    pot_eval_testing,
     nncf_optimize_testing,
     nncf_export_testing,
     nncf_eval_testing,
     nncf_eval_openvino_testing,
-    xfail_templates,
+    pot_optimize_testing,
+    pot_eval_testing
 )
 
-# Pre-train w/ 'car', 'tree' classes
+# Pre-train w/ 'car & tree' class
 args0 = {
-    '--train-ann-file': 'data/car_tree_bug/annotations/multilabel_car_tree.json',
+    '--train-ann-file': 'data/car_tree_bug/annotations/instances_car_tree.json',
     '--train-data-roots': 'data/car_tree_bug/images',
-    '--val-ann-file': 'data/car_tree_bug/annotations/multilabel_car_tree.json',
+    '--val-ann-file': 'data/car_tree_bug/annotations/instances_car_tree.json',
     '--val-data-roots': 'data/car_tree_bug/images',
-    '--test-ann-files': 'data/car_tree_bug/annotations/multilabel_car_tree.json',
+    '--test-ann-files': 'data/car_tree_bug/annotations/instances_car_tree.json',
     '--test-data-roots': 'data/car_tree_bug/images',
     '--input': 'data/car_tree_bug/images',
     'train_params': [
@@ -45,17 +43,17 @@ args0 = {
         '--learning_parameters.num_iters',
         '2',
         '--learning_parameters.batch_size',
-        '4',
+        '4'
     ]
 }
 
 # Class-Incremental learning w/ 'car', 'tree', 'bug' classes
 args = {
-    '--train-ann-file': 'data/car_tree_bug/annotations/multilabel_default.json',
+    '--train-ann-file': 'data/car_tree_bug/annotations/instances_default.json',
     '--train-data-roots': 'data/car_tree_bug/images',
-    '--val-ann-file': 'data/car_tree_bug/annotations/multilabel_default.json',
+    '--val-ann-file': 'data/car_tree_bug/annotations/instances_default.json',
     '--val-data-roots': 'data/car_tree_bug/images',
-    '--test-ann-files': 'data/car_tree_bug/annotations/multilabel_default.json',
+    '--test-ann-files': 'data/car_tree_bug/annotations/instances_default.json',
     '--test-data-roots': 'data/car_tree_bug/images',
     '--input': 'data/car_tree_bug/images',
     'train_params': [
@@ -63,25 +61,26 @@ args = {
         '--learning_parameters.num_iters',
         '2',
         '--learning_parameters.batch_size',
-        '4',
+        '4'
     ]
 }
 
-root = '/tmp/ote_cli_multilabel/'
+root = '/tmp/ote_cli/'
 ote_dir = os.getcwd()
 
-templates = Registry('external/model-preparation-algorithm').filter(task_type='CLASSIFICATION').templates
+templates = Registry(
+        'external/model-preparation-algorithm').filter(task_type='INSTANCE_SEGMENTATION').templates
 templates_ids = [template.model_template_id for template in templates]
 
 
-class TestToolsClsClsIncr:
+class TestToolsMPAInstanceSegmentation:
     @e2e_pytest_component
     def test_create_venv(self):
         work_dir, _, algo_backend_dir = get_some_vars(templates[0], root)
         create_venv(algo_backend_dir, work_dir)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_ote_train(self, template):
         ote_train_testing(template, root, ote_dir, args0)
         _, template_work_dir, _ = get_some_vars(template, root)
@@ -102,18 +101,16 @@ class TestToolsClsClsIncr:
     @e2e_pytest_component
     @pytest.mark.parametrize("template", templates, ids=templates_ids)
     def test_ote_eval_openvino(self, template):
-        ote_eval_openvino_testing(template, root, ote_dir, args, threshold=0.0)
+        ote_eval_openvino_testing(template, root, ote_dir, args, threshold=0.2)
 
     @e2e_pytest_component
     @pytest.mark.parametrize("template", templates, ids=templates_ids)
     def test_ote_demo(self, template):
-        pytest.skip("Demo for multi-label classification is not supported now.")
         ote_demo_testing(template, root, ote_dir, args)
 
     @e2e_pytest_component
     @pytest.mark.parametrize("template", templates, ids=templates_ids)
     def test_ote_demo_openvino(self, template):
-        pytest.skip("Demo for multi-label classification is not supported now.")
         ote_demo_openvino_testing(template, root, ote_dir, args)
 
     @e2e_pytest_component
@@ -129,13 +126,7 @@ class TestToolsClsClsIncr:
     @e2e_pytest_component
     @pytest.mark.parametrize("template", templates, ids=templates_ids)
     def test_ote_demo_deployment(self, template):
-        pytest.xfail("Demo for multi-label classification is not supported now.")
         ote_demo_deployment_testing(template, root, ote_dir, args)
-
-    @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
-    def test_ote_hpo(self, template):
-        ote_hpo_testing(template, root, ote_dir, args)
 
     @e2e_pytest_component
     @pytest.mark.parametrize("template", templates, ids=templates_ids)
