+++ external/model-preparation-algorithm/tests/ote_cli/test_segmentation.py
@@ -1,14 +1,16 @@
-"""Tests for MPA Class-Incremental Learning for image classification with OTE CLI"""
+"""Tests for MPA Class-Incremental Learning for semantic segmentation with OTE CLI"""
 # Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 #
 
 import os
+
 import pytest
 
 from ote_sdk.test_suite.e2e_test_system import e2e_pytest_component
 
 from ote_cli.registry import Registry
+
 from ote_cli.utils.tests import (
     create_venv,
     get_some_vars,
@@ -28,62 +30,49 @@ from ote_cli.utils.tests import (
     nncf_export_testing,
     nncf_eval_testing,
     nncf_eval_openvino_testing,
-    xfail_templates,
 )
 
-# Pre-train w/ 'intel', 'openvino' classes
-args0 = {
-    '--train-ann-file': '',
-    '--train-data-roots': 'data/text_recognition/initial_data',
-    '--val-ann-file': '',
-    '--val-data-roots': 'data/text_recognition/initial_data',
-    '--test-ann-files': '',
-    '--test-data-roots': 'data/text_recognition/initial_data',
-    '--input': 'data/text_recognition/initial_data/intel',
-    'train_params': [
-        'params',
-        '--learning_parameters.num_iters',
-        '2',
-        '--learning_parameters.batch_size',
-        '2',
-    ]
-}
 
-# Pre-train w/ 'intel', 'openvino', 'opencv' classes
 args = {
-    '--train-ann-file': '',
-    '--train-data-roots': 'data/text_recognition/IL_data',
-    '--val-ann-file': '',
-    '--val-data-roots': 'data/text_recognition/IL_data',
-    '--test-ann-files': '',
-    '--test-data-roots': 'data/text_recognition/IL_data',
-    '--input': 'data/text_recognition/IL_data/intel',
+    '--train-ann-file': 'data/segmentation/custom/annotations/training',
+    '--train-data-roots': 'data/segmentation/custom/images/training',
+    '--val-ann-file': 'data/segmentation/custom/annotations/training',
+    '--val-data-roots': 'data/segmentation/custom/images/training',
+    '--test-ann-files': 'data/segmentation/custom/annotations/training',
+    '--test-data-roots': 'data/segmentation/custom/images/training',
+    '--input': 'data/segmentation/custom/images/training',
     'train_params': [
         'params',
+        '--learning_parameters.learning_rate_fixed_iters',
+        '0',
+        '--learning_parameters.learning_rate_warmup_iters',
+        '25',
         '--learning_parameters.num_iters',
         '2',
         '--learning_parameters.batch_size',
-        '4',
+        '4'
     ]
 }
 
-root = '/tmp/ote_cli_multiclass/'
+root = '/tmp/ote_cli/'
 ote_dir = os.getcwd()
 
-templates = Registry('external/model-preparation-algorithm').filter(task_type='CLASSIFICATION').templates
+templates = Registry('external/model-preparation-algorithm').filter(task_type='SEGMENTATION').templates
 templates_ids = [template.model_template_id for template in templates]
 
 
-class TestToolsClsClsIncr:
+class TestToolsMPASegmentation:
     @e2e_pytest_component
     def test_create_venv(self):
         work_dir, _, algo_backend_dir = get_some_vars(templates[0], root)
         create_venv(algo_backend_dir, work_dir)
+        print(f'algo_backend_dir: {algo_backend_dir}')
+        print(f'work_dir: {work_dir}')
 
     @e2e_pytest_component
     @pytest.mark.parametrize("template", templates, ids=templates_ids)
     def test_ote_train(self, template):
-        ote_train_testing(template, root, ote_dir, args0)
+        ote_train_testing(template, root, ote_dir, args)
         _, template_work_dir, _ = get_some_vars(template, root)
         args1 = args.copy()
         args1['--load-weights'] = f'{template_work_dir}/trained_{template.model_template_id}/weights.pth'
@@ -102,7 +91,7 @@ class TestToolsClsClsIncr:
     @e2e_pytest_component
     @pytest.mark.parametrize("template", templates, ids=templates_ids)
     def test_ote_eval_openvino(self, template):
-        ote_eval_openvino_testing(template, root, ote_dir, args, threshold=0.0)
+        ote_eval_openvino_testing(template, root, ote_dir, args, threshold=0.1)
 
     @e2e_pytest_component
     @pytest.mark.parametrize("template", templates, ids=templates_ids)
@@ -169,9 +158,13 @@ class TestToolsClsClsIncr:
     @e2e_pytest_component
     @pytest.mark.parametrize("template", templates, ids=templates_ids)
     def test_pot_optimize(self, template):
+        if template.model_template_id.startswith('ClassIncremental_Semantic_Segmentation_Lite-HRNet-'):
+            pytest.skip('CVS-82482')
         pot_optimize_testing(template, root, ote_dir, args)
 
     @e2e_pytest_component
     @pytest.mark.parametrize("template", templates, ids=templates_ids)
     def test_pot_eval(self, template):
+        if template.model_template_id.startswith('ClassIncremental_Semantic_Segmentation_Lite-HRNet-'):
+            pytest.skip('CVS-82482')
         pot_eval_testing(template, root, ote_dir, args)
