--- tests/ote_cli/external/deep-object-reid/ote_cli
@@ -1,1 +1,0 @@
-../../../../external/deep-object-reid/tests/ote_cli
