--- tests/ote_cli/external/mmdetection/ote_cli
@@ -1,1 +1,0 @@
-../../../../external/mmdetection/tests/ote_cli
