--- tests/ote_cli/external/mmsegmentation/ote_cli
@@ -1,1 +1,0 @@
-../../../../external/mmsegmentation/tests/ote_cli
